pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tailsocks"
version = "0.1.0"
description = "A Tailscale SOCKS5 proxy manager"
readme = "README.md"
authors = [
    {name = "James Dixson", email = "james@yoshikostudios.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Networking",
]
requires-python = ">=3.7"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "attrs>=23.1.0",
]
dev = [
    "ruff>=0.1.6",
]

[project.scripts]
tailsocks = "tailsocks.cli:main"

[project.urls]
"Homepage" = "https://github.com/yoshikostudios/tailsocks"
"Bug Tracker" = "https://github.com/yoshikostudios/tailsocks/issues"

[tool.ruff]
# Enable flake8-bugbear (`B`) rules.
select = ["E", "F", "B", "I", "W", "C90"]
# Ignore line length errors
ignore = ["E501"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.7
target-version = "py37"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.isort]
known-first-party = ["tailsocks"]

[tool.ruff.format]
# Use double quotes for strings.
quote-style = "double"
# Indent with 4 spaces.
indent-style = "space"
line-ending = "auto"
